changes made to code

+++ mini_project.py
@@ -12,55 +12,57 @@ class Classfirst:
 class Searchclass(Classfirst):  # class to perform search operation
 
     def function_search_writefile(self):
+        if self.nos.isdigit():
 
-        for _ in range(int(self.nos)):
+            for _ in range(int(self.nos)):
 
-            smt = ''
+                smt = ''
 
-            count = 0
+                count = 0
 
-            Search_key = input("Enter Search Key\n")
+                Search_key = input("Enter Search Key\n")
 
-            file_name = Search_key + '.txt'
+                file_name = Search_key + '.txt'
 
-            write_file = open(file_name, 'w')
+                write_file = open(file_name, 'w')
 
-            read_file = open("Input.txt", 'r')
+                read_file = open("Input.txt", 'r')
 
-            read_file1 = read_file.readlines()
+                read_file1 = read_file.readlines()
 
-            for line in read_file1:
-                smt += line
+                for line in read_file1:
+                    smt += line
 
-            smt = re.sub('\W+', ' ', smt)
+                smt = re.sub(r'\W+', ' ', smt)
 
-            newstr = smt.split(' ')
+                newstr = smt.split(' ')
 
-            for x in range(len(newstr)):
+                for x in range(len(newstr)):
 
-                if newstr[x].lower() == Search_key.lower():
-                    write_file.write(newstr[x - 1] + ' ' + newstr[x] + ' ' +
-                                     newstr[x + 1] + '\n')
+                    if newstr[x].lower() == Search_key.lower():
+                        write_file.write(newstr[x - 1] + ' ' + newstr[x] +
+                                         ' ' + newstr[x + 1] + '\n')
 
-                    count += 1
+                        count += 1
 
-            if count != 0:
-                write_file.write('Total count of ' + Search_key +
-                                 ' in input file is ' + str(count))
+                if count != 0:
+                    write_file.write('Total count of ' + Search_key +
+                                     ' in input file is ' + str(count))
 
-            write_file.close()
+                write_file.close()
 
-            if count == 0:
+                if count == 0:
 
-                print("Keyword not found in the file\n")
+                    print("Keyword not found in the file\n")
 
-                if os.path.isfile(file_name):
-                    os.remove(file_name)
+                    if os.path.isfile(file_name):
+                        os.remove(file_name)
 
-            write_file.close()
+                write_file.close()
 
-            read_file.close()
-
+                read_file.close()
+        else:
+            print("Please enter the correct input value")
 
 clas_object = Searchclass()
